Move CDMNode and CDMSort to dap4 module.
