+++ dap4/d4cdm/src/main/java/dap4/cdm/CDMSort.java
@@ -1,9 +1,9 @@
 /*
- * Copyright (c) 1998-2018 John Caron and University Corporation for Atmospheric Research/Unidata
- * See LICENSE for license information.
+ * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
+ *  See LICENSE for license information.
  */
 
-package ucar.nc2;
+package dap4.cdm;
 
 /**
  * Enumeration of the kinds of nodes in a CDM model.
